requirements.txt
infi.systray
mouse
keyboard

# Building requires:
PyInstall